+++ Example/GBHFacebookImagePicker/ViewController.swift
@@ -84,12 +84,6 @@ class ViewController: UIViewController {
         // Album's count color 
         GBHFacebookImagePicker.pickerConfig.uiConfig.albumsCountColor = UIColor.white
 
-        // Selected border color 
-        GBHFacebookImagePicker.pickerConfig.uiConfig.selectedBorderColor = UIColor.red
-
-        // Selected border width 
-        GBHFacebookImagePicker.pickerConfig.uiConfig.selectedBorderWidth = 4.0
-
         // Maximum selected pictures 
         GBHFacebookImagePicker.pickerConfig.maximumSelectedPictures = 4
 
@@ -99,8 +93,8 @@ class ViewController: UIViewController {
         // Tagged album name
         GBHFacebookImagePicker.pickerConfig.displayTaggedAlbum = true
 
-        /// Number of picture per row (3 by default)
-        GBHFacebookImagePicker.pickerConfig.picturePerRow = 4
+        /// Number of picture per row (4 by default)
+        GBHFacebookImagePicker.pickerConfig.picturePerRow = 3
         
         /// Space beetween album photo cell (1.5 by default)
         GBHFacebookImagePicker.pickerConfig.cellSpacing = 2.0