+++ GBHFacebookImagePicker/Classes/Model/GBHAssetImage.swift
@@ -1,0 +1,17 @@
+//
+//  GBHAssetImage.swift
+//  Bolts
+//
+//  Created by Florian Gabach on 10/09/2017.
+//
+
+import Foundation
+
+/// Struct to define assets name 
+struct GBHAssetImage {
+    /// Default image loader name
+    static let loader = "GBHFacebookImagePickerDefaultImageLoading"
+
+    /// Checkmark logo name
+    static let checkMark = "GBHFacebookImagePickerCheckMark"
+}